fix: allow kdump initrd regeneration and fix path
There are two issues. First, our default kdump configuration causes the
kdump service to use the host's default initrd as the crashkernel
environment. This can lead to issues when the default initrd size is
larger than the reserved memory for the crash kernel, which is set via
kernel command line at boot time.

It is common to have the kdump service instead rebuild a minimal
host-specific initrd specifically for the crashkernel environment.

To change this behavior, comment out force_no_rebuild 1 from kdump.conf.

After doing this, a second issue was discovered where the system would
enter the crashkernel environment when a kernel panic occurred, and
successfully collect the crash dump, but upon reboot back to the normal
host OS, the system would no longer boot.

This behavior was root caused to an issue in our kdumpctl command where
the TARGET_INITRD was pointing to our default host initrd, thus when the
kdump service would regenerate the minimal host-specific initrd, this
new initrd would overwrite the host's normal initrd, therefore leading
to failed normal boot.

To fix this, change the TARGET_INITRD path to be unique per kernel
version.

Signed-off-by: Chris Co <chrco@microsoft.com>