fix: Add requires libxcrypt to glibc package
In 2.38, glibc no longer builds libcrypt.so by default. We also do not
specify building libcrypt in the configure stage. This is desired, since
we want to use libxcrypt's version of libcrypt instead. libxcrypt is
a more modern version of the crypt library. So in glibc, add a package
requires to pull in libxcrypt for all installs of glibc

Signed-off-by: Chris Co <chrco@microsoft.com>